Refactoring errors
Previously the errors were returned
when the secret key was generated.
The return of the secret key is now
a struct so this is removed.

The changes to the error file include
a rewriting of the description behind
the generic error. As well as inserting
a new error type for when a signature
cannot be verified properly.